WIP: switch encoding unsigned type normalization to Uint* types
Using Java native types for unsigned values is wasteful especially
in case of uint32 and uint64. Furthermore it leads to uncertainty
between in16/uint8, int32/uint16 and int64/uint32.

This is a partial switch, where we retain java.lang types for
signed entities, but switch the unsigned ones.

JIRA: YANGTOOLS-1018

+++ yang/yang-model-api/src/main/java/org/opendaylight/yangtools/yang/model/api/type/Uint8TypeDefinition.java
@@ -9,15 +9,14 @@ package org.opendaylight.yangtools.yang.model.api.type;
 
 import org.eclipse.jdt.annotation.NonNull;
 import org.eclipse.jdt.annotation.Nullable;
+import org.opendaylight.yangtools.yang.common.Uint8;
 
 /**
  * Type definition derived from uint8 type.
  *
  * @author Robert Varga
  */
-public interface Uint8TypeDefinition extends RangeRestrictedTypeDefinition<Uint8TypeDefinition, Short> {
-
-
+public interface Uint8TypeDefinition extends RangeRestrictedTypeDefinition<Uint8TypeDefinition, Uint8> {
     static int hashCode(final @NonNull Uint8TypeDefinition type) {
         return TypeDefinitions.hashCode(type);
     }
